requirements.txt
loguru==0.6.0
requests>=2.26.0
tabulate==0.8.9
wcwidth==0.2.5
flask
pydantic>=2.6.0
python-dotenv
curl_cffi
redis>=4.0.0
pymysql>=1.0.0

# LangChain for AI analysis
langchain>=0.3.27
langchain-openai>=0.3.35
langchain-core>=0.3.80

# DuckDuckGo search for fact verification
ddgs>=1.0.0

# Web scraping for fact verification
beautifulsoup4>=4.12.0
lxml>=5.0.0

urllib3
akshare>=1.14.0
